requirements.txt
a2wsgi==1.10.10
aiosqlite==0.21.0
alembic==1.16.4
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
appnope==0.1.4
argcomplete==3.6.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.9.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
awsebcli==3.25
babel==2.16.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
bleach==6.1.0
blessed==1.22.0
blinker==1.9.0
botocore==1.40.38
build==1.2.2.post1
cachetools==5.5.0
cadwyn==5.4.4
cement==2.10.14
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==45.0.6
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
distlib==0.3.8
dnspython==2.7.0
email_validator==2.2.0
env==0.1.0
executing==2.1.0
fabric==3.2.2
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fastjsonschema==2.20.0
filelock==3.15.4
Flask==3.1.1
fqdn==1.5.1
frozendict==2.4.6
fsspec==2025.7.0
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.7
importlib_metadata==8.7.0
invoke==2.2.0
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.6.12
langchain-core==0.3.9
langgraph==0.2.34
langgraph-checkpoint==2.0.1
langsmith==0.1.132
lazy-object-proxy==1.11.0
libcst==1.8.2
linkify-it-py==2.0.3
lockfile==0.12.2
longchain==0.4.2
lxml==5.3.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
methodtools==0.4.7
mistune==3.0.2
more-itertools==10.7.0
msgpack==1.1.0
msgspec==0.19.0
multitasking==0.0.11
narwhals==1.8.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.2.4
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.10.7
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
paramiko==4.0.0
parso==0.8.4
pathspec==0.12.1
peewee==3.17.7
pendulum==3.1.0
pexpect==4.9.0
pillow==10.4.0
pip-tools==7.4.1
platformdirs==4.2.2
pluggy==1.6.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==24.12.1
pyarrow==17.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
PyNaCl==1.6.0
pyproject_hooks==1.2.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==2.0.7
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.31.0
requests-toolbelt==1.0.0
retryhttp==1.3.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rich-argparse==1.7.1
rich-toolkit==0.14.9
rignore==0.6.4
rpds-py==0.20.0
scikit-learn==1.6.1
scikit-optimize==0.10.2
scipy==1.14.1
semantic-version==2.10.0
Send2Trash==1.8.3
sentry-sdk==2.34.1
setproctitle==1.3.6
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.2
structlog==25.4.0
svcs==25.1.0
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.5.0
terminado==0.18.1
text-unidecode==1.3
textdistance==4.6.1
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
typer==0.16.0
types-python-dateutil==2.9.0.20241003
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
universal_pathlib==0.2.6
uri-template==1.3.0
urllib3==1.26.20
uuid6==2025.0.1
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.26.3
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
wirerope==1.0.0
wrapt==1.17.2
yfinance==0.2.44
zipp==3.23.0
